finish refactor download dialog
