+++ app/src/main/java/com/claire/traveldiary/settings/download/DownloadContract.java
@@ -7,11 +7,15 @@ public interface DownloadContract {
 
     interface View extends BaseView<Presenter> {
 
+        void successfullyDownloadUi();
+
     }
 
     interface Presenter extends BasePresenter {
 
         void result(int requestCode, int resultCode);
+
+        void ReadDataFromFirebase();
 
     }
 }
